feat: add Cyan color shades and refactor Green color shades
- Add Cyan color shades with hex and RGB conversion methods
- Refactor Green color shades to remove redundant attributes
- Update module exports to include Cyan color
